Functions for setting and pruning empty properties
`assign_empty_defaults` intends to enable the simplification of logic that
currently depends on the `(collection[value] || [])` idiom to avoid `nil`
values.

`prune_empty_properties` intends to undo this by replacing empty values with
`nil`. In the Liquid templates used by Jekyll, this enables the use of
expressions such as `{% if collection.value %}` rather than
`{% if collection.value != empty %}` or
`{% if collection.value.size != 0 %}`.